\name{readgoogle}
\alias{readgoogle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
readgoogle
}
\description{
Read in data directly from google sheet.
}
\usage{
readgoogle(url)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{url}{
a web link
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{url }{ the web link of your google sheet}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

